move RecordReader interface to sql.execution.datasources
