+++ src/main/java/org/apache/parquet/hadoop/MrOapRecordReader.java
@@ -23,7 +23,7 @@ import java.io.IOException;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
 import org.apache.parquet.hadoop.api.ReadSupport;
-import org.apache.parquet.hadoop.api.RecordReader;
+import org.apache.spark.sql.execution.datasources.RecordReader;
 import org.apache.parquet.hadoop.metadata.ParquetFooter;
 
 import static org.apache.parquet.hadoop.ParquetInputFormat.getFilter;
