+++ src/main/scala/org/apache/spark/sql/execution/datasources/oap/io/VectorizedCacheReader.scala
@@ -22,7 +22,7 @@ import java.io.IOException
 import scala.collection.JavaConverters._
 
 import org.apache.hadoop.conf.Configuration
-import org.apache.parquet.hadoop.api.{InitContext, RecordReader}
+import org.apache.parquet.hadoop.api.InitContext
 import org.apache.parquet.hadoop.metadata._
 import org.apache.parquet.hadoop.utils.Collections3
 import org.apache.parquet.schema.{MessageType, Type}
@@ -32,6 +32,7 @@ import org.apache.spark.memory.MemoryMode
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.execution.datasources.oap.filecache.DataFiberId
 import org.apache.spark.sql.execution.datasources.parquet.ParquetReadSupportWrapper
+import org.apache.spark.sql.execution.datasources.RecordReader
 import org.apache.spark.sql.execution.vectorized._
 import org.apache.spark.sql.oap.OapRuntime
 import org.apache.spark.sql.types.StructType
